requirements.txt
docopt==0.6.2
Jinja2==2.8
requests==2.20.0
BeautifulSoup==3.2.1
Pillow==4.0.0
beautifulsoup4==4.5.3
